Updated cart page
Fixed issues on cart page with "drift up" when cart item quantities were changed. This also required us to totally rebuild the “YOUR CART/QUOTE” header text with the number of items in your cart because that was causing a ton of issues because it’s updated server-side.

These changes also implemented the new "YOU SAVED" widget on the cart page with yet-to-be-confirmed exact copy verbiage.